requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.8.2
python-multipart==0.0.9
